:construction: Ⓢ WIP: Add Singularity Support
- :sparkles: Pull image
- :sparkles: Load image
- :sparkles: Run image

on-behalf-of: @FCP-INDI <http://fcp-indi.github.com>
